requirements.txt
deepface>=0.0.92
opencv-python>=4.10.0.84
tensorflow-macos>=2.16.1
tensorflow-metal>=1.1.0
tf-keras>=2.16.0
pillow>=10.0.0
numpy>=1.26.0
